int -> long long

+++ ft_printf.h
@@ -22,14 +22,14 @@
 # include <stdio.h>
 
 int				ft_printf(const char *format, ...);
-int				ft_putchar(char c, int count);
-int				ft_putstr(char *s, int count);
-int				ft_putnbr(long args, int count);
-unsigned int	ft_put_conv_base(unsigned long long args, char x_X, int count);
-int				if_d_or_i(va_list *args, int count);
-int				if_s(va_list *args, int count);
-int				if_c(va_list *args, int count);
-int				if_u(va_list *args, int count);
-int				if_p_x(va_list *args, int count, char c);
+long long		ft_putchar(char c, long long count);
+long long		ft_putstr(char *s, long long count);
+long long		ft_putnbr(long args, long long count);
+long long		ft_put_conv_base(unsigned long long args, char x_X, long long count);
+long long		if_d_or_i(va_list *args, long long count);
+long long		if_s(va_list *args, long long count);
+long long		if_c(va_list *args, long long count);
+long long		if_u(va_list *args, long long count);
+long long		if_p_x(va_list *args, long long count, char c);
 
 #endif
